Move plugin lister to avoid import cycle.
